+++ MoveClient/MoveData.cs
@@ -13,18 +13,18 @@ namespace MoveServer
         public int Id { get; set; }
 
         [Interpolation]
-        public double X { get; set; }
+        public float X { get; set; }
         [Interpolation]
-        public double Y { get; set; }
-        public double VelX { get; set; }
-        public double VelY { get; set; }
+        public float Y { get; set; }
+        public float VelX { get; set; }
+        public float VelY { get; set; }
 
         public MoveData() : this(0,0,0,0)
         {
 
         }
 
-        public MoveData(double x, double y, double velx, double vely)
+        public MoveData(float x, float y, float velx, float vely)
         {
             X = x;
             Y = y;
